requirements.txt
flask==3.0.3
Flask-RESTful==0.3.10
pydantic==2.7.3
PyYAML==6.0.1
requests==2.32.3
easydict==1.13
asyncio==3.4.3
aiohttp==3.9.5
psycopg2-binary==2.9.9